+++ sdk/python/nemo-platform/api.md
@@ -10,6 +10,8 @@ from nemo_platform.types import (
     DatetimeFilter,
     DeleteResponse,
     FileStorageType,
+    FilesetMetadata,
+    FilesetMetadataParam,
     FinetuningType,
     GenericSortField,
     HTTPValidationError,
