+++ sdk/python/nemo-platform/src/nemo_platform/resources/files/api.md
@@ -33,7 +33,7 @@ Methods:
 Types:
 
 ```python
-from nemo_platform.types.files import FilesetFilter, FilesetMetadata, FilesetMetadataParam
+from nemo_platform.types.files import FilesetFilter
 ```
 
 Methods:
