+++ sdk/python/nemo-platform/src/nemo_platform/types/__init__.py
@@ -32,6 +32,7 @@ from .shared import (
     PlatformJobLog as PlatformJobLog,
     ToolCallConfig as ToolCallConfig,
     APIEndpointData as APIEndpointData,
+    FilesetMetadata as FilesetMetadata,
     FileStorageType as FileStorageType,
     InferenceParams as InferenceParams,
     LinearLayerSpec as LinearLayerSpec,
@@ -41,6 +42,7 @@ from .shared import (
     PlatformJobLogPage as PlatformJobLogPage,
     HTTPValidationError as HTTPValidationError,
     SlidingWindowConfig as SlidingWindowConfig,
+    FilesetMetadataParam as FilesetMetadataParam,
     ModelMetadataContent as ModelMetadataContent,
     AuthDiscoveryResponse as AuthDiscoveryResponse,
     OidcDiscoveryResponse as OidcDiscoveryResponse,
