+++ sdk/python/nemo-platform/src/nemo_platform/types/files/__init__.py
@@ -22,7 +22,6 @@ from .secret_ref import SecretRef as SecretRef
 from .cache_status import CacheStatus as CacheStatus
 from .fileset_file import FilesetFile as FilesetFile
 from .fileset_purpose import FilesetPurpose as FilesetPurpose
-from .fileset_metadata import FilesetMetadata as FilesetMetadata
 from .s3_storage_config import S3StorageConfig as S3StorageConfig
 from .ngc_storage_config import NGCStorageConfig as NGCStorageConfig
 from .fileset_list_params import FilesetListParams as FilesetListParams
@@ -33,7 +32,6 @@ from .local_storage_config import LocalStorageConfig as LocalStorageConfig
 from .fileset_create_params import FilesetCreateParams as FilesetCreateParams
 from .fileset_update_params import FilesetUpdateParams as FilesetUpdateParams
 from .file_list_files_params import FileListFilesParams as FileListFilesParams
-from .fileset_metadata_param import FilesetMetadataParam as FilesetMetadataParam
 from .file_upload_file_params import FileUploadFileParams as FileUploadFileParams
 from .s3_storage_config_param import S3StorageConfigParam as S3StorageConfigParam
 from .ngc_storage_config_param import NGCStorageConfigParam as NGCStorageConfigParam
