+++ sdk/python/nemo-platform/src/nemo_platform/types/shared_params/__init__.py
@@ -31,6 +31,7 @@ from .linear_layer_spec import LinearLayerSpec as LinearLayerSpec
 from .generic_sort_field import GenericSortField as GenericSortField
 from .platform_job_status import PlatformJobStatus as PlatformJobStatus
 from .sliding_window_config import SlidingWindowConfig as SlidingWindowConfig
+from .fileset_metadata_param import FilesetMetadataParam as FilesetMetadataParam
 from .model_metadata_content import ModelMetadataContent as ModelMetadataContent
 from .dataset_metadata_content import DatasetMetadataContent as DatasetMetadataContent
 from .tool_calling_metadata_content import ToolCallingMetadataContent as ToolCallingMetadataContent
